+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Convert response message targets
+
 ## Version 1.5.5 - 2021-01-27
 
 ### Fixed